backend/chatbotbf/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
httpx==0.28.1